- Removed multiple files related to vendor/ImGui - Removed Android examples and documentation files in ImGui
The commit removed multiple files related to vendor/ImGui. This includes Android examples and documentation files under the ImGui directory. The decision to remove these files came from the necessity to clean up the codebase and remove unnecessary or unused files.

+++ src/Ayla/WindowSystem/Window.h
@@ -16,8 +16,8 @@ namespace Ayla{
 */
     struct WindowProperties {
         std::string title; /**< The title of the window */
-        int width; /**< The width of the window in pixels */
-        int height; /**< The height of the window in pixels */
+        float width; /**< The width of the window in pixels */
+        float height; /**< The height of the window in pixels */
 
         /**
          * @brief Constructor for WindowProperties
